requirementsall.txt
altgraph==0.16.1
awscli==1.16.280
boto3==1.10.16
botocore==1.13.16
circle-fit==0.1.3
Click==7.0
colorama==0.4.1
cycler==0.10.0
docutils==0.15.2
dominate==2.4.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.1
opencv-python==4.1.1.26
Pillow==6.2.0
pyasn1==0.4.7
PyInstaller==3.5
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
redis==3.3.11
retrying==1.3.3
rq==1.1.0
rsa==3.4.2
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
